+++ packages/terra-layout/src/terra-dev-site/test/common/TestHeader.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames/bind';
+import styles from './TestHeader.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
   layoutConfig: PropTypes.object,
@@ -10,23 +14,15 @@ const TestHeader = ({
 }) => (
   <div
     id="test-header"
-    style={{
-      height: '60px', width: 'calc(100% - 10px)', border: '4px dashed lightgrey', margin: '5px', position: 'relative', display: 'flex',
-    }}
+    className={cx('content-wrapper1')}
   >
-    <div style={{
-      position: 'absolute', top: '15px', left: '50%', color: 'grey', transform: 'translateX(-50%)',
-    }}
-    >
-      <h2 style={{ display: 'inline', margin: 0 }}>Header</h2>
+    <div className={cx('content-wrapper2')}>
+      <h2 className={cx('content-text')}>Header</h2>
       {layoutConfig.toggleMenu
         && (
         <button
           type="button"
-          className="test-header-toggle"
-          style={{
-            display: 'inline', marginLeft: '5px', height: '25px', border: '1px dashed lightgrey',
-          }}
+          className={cx('test-header-toggle')}
           onClick={layoutConfig.toggleMenu}
         >
 Toggle Menu