+++ packages/terra-time-input/tests/wdio/time-input-twelve-hour-mobile-spec.js
@@ -1,13 +1,10 @@
-const viewports = Terra.viewports('medium');
-
 // Color contrast will be resolved in https://github.com/cerner/terra-core/issues/1670
 const ignoredA11y = {
   'color-contrast': { enabled: false },
 };
 
-describe('Time Input Twelve Hour Mobile', () => {
+Terra.describeViewports('Time Input Twelve Hour Mobile', ['medium'], () => {
   before(() => {
-    browser.setViewportSize(Terra.viewports('tiny')[0]);
     browser.execute('if (!window.ontouchstart) { window.ontouchstart = "true"; }');
   });
 
@@ -18,8 +15,8 @@ describe('Time Input Twelve Hour Mobile', () => {
   describe('Twelve Hour Mobile - Default with no time', () => {
     before(() => browser.url('/#/raw/tests/terra-time-input/time-input/twelve-hour/mobile'));
 
-    Terra.it.matchesScreenshot({ viewports });
-    Terra.it.isAccessible({ viewports, rules: ignoredA11y });
+    Terra.it.matchesScreenshot();
+    Terra.it.isAccessible({ rules: ignoredA11y });
   });
 
   describe('Twelve Hour Mobile - Clicking Meridiem updates the time', () => {
@@ -29,7 +26,7 @@ describe('Time Input Twelve Hour Mobile', () => {
       browser.click('#timeInputValueProvided button:not([aria-pressed="true"])');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Mobile - Focus Hour Styles', () => {
@@ -45,8 +42,8 @@ describe('Time Input Twelve Hour Mobile', () => {
       browser.click('#timeInput input[name="terra-time-hour-time-input"]');
     });
 
-    Terra.it.isAccessible({ viewports, rules: ignoredA11y });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.isAccessible({ rules: ignoredA11y });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Mobile - Focus Minute Styles', () => {
@@ -62,8 +59,8 @@ describe('Time Input Twelve Hour Mobile', () => {
       browser.click('#timeInput input[name="terra-time-minute-time-input"]');
     });
 
-    Terra.it.isAccessible({ viewports, rules: ignoredA11y });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.isAccessible({ rules: ignoredA11y });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Mobile - Focus Second Styles', () => {