+++ packages/terra-time-input/tests/wdio/time-input-twelve-hour-spec.js
@@ -1,24 +1,20 @@
-const viewports = Terra.viewports('medium');
-
-describe('Time Input Twelve Hour', () => {
-  before(() => browser.setViewportSize(Terra.viewports('tiny')[0]));
-
+Terra.describeViewports('Time Input Twelve Hour', ['medium'], () => {
   describe('Twelve Hour Default with no time provided', () => {
     before(() => browser.url('/#/raw/tests/terra-time-input/time-input/twelve-hour/default'));
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Default with an evening time filled', () => {
     before(() => browser.url('/#/raw/tests/terra-time-input/time-input/twelve-hour/filled-evening'));
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Default with a morning time filled', () => {
     before(() => browser.url('/#/raw/tests/terra-time-input/time-input/twelve-hour/filled-morning'));
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Up switches to a.m.', () => {
@@ -29,7 +25,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowUp']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Down switches to p.m.', () => {
@@ -41,7 +37,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowDown']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Auto focuses on meridiem when filling out whole time', () => {
@@ -58,7 +54,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys('1234p');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Auto focuses on meridiem when filling out minute', () => {
@@ -75,7 +71,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys('34p');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Sets hour to 12 without meridiem change for hour input 00', () => {
@@ -92,7 +88,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys('00');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Sets hour to 12 without meridiem change for hour input 00', () => {
@@ -109,7 +105,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys('00');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Up Arrow on hour does not change meridiem', () => {
@@ -126,7 +122,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowUp']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Changes time to 01 when up is pressed on hour of 12', () => {
@@ -145,7 +141,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowUp']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Changes time to 12 when down is pressed on hour of 01', () => {
@@ -164,7 +160,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowDown']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Switched when up is press on hour of 11', () => {
@@ -183,7 +179,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowUp']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Meridiem - Switched when down is press on hour of 12', () => {
@@ -202,7 +198,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowDown']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Sets time to 12 when hour is 0 and onBlur is called', () => {
@@ -220,7 +216,7 @@ describe('Time Input Twelve Hour', () => {
       browser.click('#timeInput input[name="terra-time-minute-time-input"]');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Sets time to 12 when hour is 0 and onBlur is called', () => {
@@ -238,7 +234,7 @@ describe('Time Input Twelve Hour', () => {
       browser.click('#timeInput input[name="terra-time-minute-time-input"]');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Sets time to 12 when hour is 0 and onBlur is called', () => {
@@ -256,7 +252,7 @@ describe('Time Input Twelve Hour', () => {
       browser.click('#timeInput input[name="terra-time-minute-time-input"]');
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Focuses to minute when left is pressed from meridiem', () => {
@@ -273,7 +269,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['ArrowLeft']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour - Focuses to minute when delete is pressed from meridiem', () => {
@@ -290,7 +286,7 @@ describe('Time Input Twelve Hour', () => {
       browser.keys(['Delete']);
     });
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.matchesScreenshot();
   });
 
   describe('Twelve Hour Default with second and no time provided', () => {