+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Added
 * Added the ability for the user to enter seconds in the time input
 
+4.4.0 - (June 19, 2019)
+------------------
+### Changed
+* Update wdio specs to use Terra.describeViewports
+
 4.3.0 - (June 12, 2019)
 ------------------
 ### Changed