+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+4.13.0 - (June 19, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.12.0 - (June 12, 2019)
 ------------------s
 ### Removed