+++ packages/terra-popup/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+6.3.0 - (June 19, 2019)
+------------------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
 6.2.0 - (June 12, 2019)
 ------------------
 ### Removed
@@ -11,6 +16,7 @@ Unreleased
 
 ### Changed
 * Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+* Update wdio specs to use Terra.describeViewports
 
 6.1.0 - (May 21, 2019)
 ------------------